I added two new validations, added two unit tests and updated logs and rename the validation.py to menu_validation.py
