+++ src/com/jalasoft/search_files/menu/cli_menu_advanced_flow.py
@@ -1,5 +1,5 @@
 from src.com.jalasoft.search_files.menu.cli_menu import CLIMenu
-from src.com.jalasoft.search_files.utils.validation import Validation
+from src.com.jalasoft.search_files.utils.menu_validation import Validation
 from src.com.jalasoft.search_files.search.search import Search
 from src.com.jalasoft.search_files.search.file import File
 from src.com.jalasoft.search_files.search.directory import Directory
