+++ src/com/jalasoft/search_files/utils/menu_validation.py
@@ -12,7 +12,7 @@ class Validation(object):
         logger.info("Starting the method")
         boolean_contain = string_value.isdigit()
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_the_range(self, number, upper_limit, lower_limit):
@@ -23,7 +23,7 @@ class Validation(object):
         else:
             boolean_contain = False
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def attempt_maximum(self, attempts):
@@ -34,7 +34,7 @@ class Validation(object):
         else:
             boolean_contain = False
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_path_valid(self, directory):
@@ -42,7 +42,7 @@ class Validation(object):
         directory = str(directory)
         boolean_contain = os.path.exists(directory)
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def has_valid_characters(self, string_value):
@@ -50,7 +50,7 @@ class Validation(object):
         string_value = str(string_value)
         boolean_contain = string_value.isalnum()
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_blank(self, string_value):
@@ -61,7 +61,7 @@ class Validation(object):
         else:
             boolean_contain = True
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_valid_date(self, str_date):
@@ -78,12 +78,13 @@ class Validation(object):
                 datetime.datetime.strptime(str_date, date_format)
             except ValueError:
                 logger.info("The value returned is: %s", False)
+                logger.info("Ending the method")
                 return False
             boolean_contain = True
         else:
             boolean_contain = False
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
