+++ src/com/jalasoft/search_files/utils/search_validation.py
@@ -1,5 +1,6 @@
 import re
 import bitmath
+import sys
 from src.com.jalasoft.search_files.utils.logging_config import logger
 
 
@@ -23,7 +24,7 @@ class SearchValidation(object):
             else:
                 boolean_contain = True
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_string1_equal_to_string2(self, string_1, string_2, flag):
@@ -42,7 +43,7 @@ class SearchValidation(object):
             else:
                 boolean_contain = True
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_less_than_or_equal(self, bit_size, mega_bit_size):
@@ -54,7 +55,7 @@ class SearchValidation(object):
         else:
             boolean_contain = False
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
 
     def is_bigger_than_other(self, bit_size, mega_bit_size):
@@ -67,5 +68,35 @@ class SearchValidation(object):
         else:
             boolean_contain = False
         logger.info("The value returned is: %s", boolean_contain)
-        logger.info("Exit method")
+        logger.info("Ending the method")
         return boolean_contain
+
+    def is_windows(self):
+        logger.info("Starting the method")
+        operation_system = sys.platform
+        search_match = re.search("win", operation_system)
+        if search_match is None:
+            boolean_contain = False
+        else:
+            boolean_contain = True
+        logger.info("The value returned is: %s", boolean_contain)
+        logger.info("Ending the method")
+        return boolean_contain
+
+    def is_linux(self):
+        logger.info("Starting the method")
+        operation_system = sys.platform
+        search_match = re.search("linux", operation_system)
+        if search_match is None:
+            boolean_contain = False
+        else:
+            boolean_contain = True
+        logger.info("The value returned is: %s", boolean_contain)
+        logger.info("Ending the method")
+        return boolean_contain
+
+
+
+
+
+
